refactor(ui): extract cva variants for Fast Refresh
Move the badge, button, and tabs cva() definitions into dedicated
*-variants.ts modules so the component files export only components.
This resolves the react-refresh/only-export-components lint warnings
without an eslint-disable. No external code imported these helpers from
the component path, so the variant exports are dropped rather than
re-exported.
